Refactor package from todo to todolistSample
