+++ finish/src/main/java/io/openliberty/guides/todolistSample/TodoApplication.java
@@ -1,4 +1,4 @@
-package io.openliberty.guides.todo;
+package io.openliberty.guides.todolistSample;
 
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
