+++ finish/src/main/java/io/openliberty/guides/todolistSample/resources/TodoResource.java
@@ -1,4 +1,4 @@
-package io.openliberty.guides.todo.resources;
+package io.openliberty.guides.todolistSample.resources;
 
 import java.util.Optional;
 
@@ -15,8 +15,8 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
-import io.openliberty.guides.todo.models.TodoModel;
-import io.openliberty.guides.todo.services.TodoService;
+import io.openliberty.guides.todolistSample.models.*;
+import io.openliberty.guides.todolistSample.services.*;
 
 @Path("todos")
 public class TodoResource {
