+++ finish/src/main/resources/META-INF/beans.xml
@@ -1,5 +1,5 @@
 <beans>
     <alternatives>
-        <class>io.openliberty.guides.todo.services.samples.SampleTodoService</class>
+        <class>io.openliberty.guides.todolistSample.services.samples.SampleTodoService</class>
     </alternatives>
 </beans>
